refactor(root): delete logger package and move it to utils
