+++ README.md
@@ -176,7 +176,6 @@ The monorepo uses [Renovate](https://docs.renovatebot.com) for intelligent depen
 | `api` | App | [Express](https://expressjs.com/en/guide/routing.html) + [TypeScript](https://www.typescriptlang.org/docs/) | 3003 | Backend API server |
 | `@repo/ui` | Package | [React](https://react.dev/learn) | 3004 | Shared UI components with Storybook |
 | `@repo/utils` | Package | [TypeScript](https://www.typescriptlang.org/docs/) | - | Utility functions |
-| `@repo/logger` | Package | [TypeScript](https://www.typescriptlang.org/docs/) | - | Logging utilities |
 | `@repo/test-preset` | Package | [Jest](https://jestjs.io/docs/getting-started) | - | Testing configurations |
 | `@repo/config-typescript` | Package | [TypeScript](https://www.typescriptlang.org/docs/) | - | TypeScript configurations |
