+++ apps/api/package.json
@@ -10,7 +10,7 @@
 		"check:types": "tsc --noEmit"
 	},
 	"dependencies": {
-		"@repo/logger": "workspace:*",
+		"@repo/utils": "workspace:*",
 		"body-parser": "2.2.0",
 		"cors": "2.8.5",
 		"express": "5.1.0",
