+++ apps/api/src/index.ts
@@ -1,4 +1,4 @@
-import { log } from "@repo/logger";
+import { log } from "@repo/utils/logger";
 import { createServer } from "./server";
 
 const host = process.env.HOST || "localhost";
