--- packages/logger/CHANGELOG.md
@@ -1,7 +1,0 @@
-# @repo/logger
-
-## 0.0.1
-
-### Patch Changes
-
-- Automated version bump for storefront and other affected packages on root version tag: v0.0.1.
