--- packages/logger/package.json
@@ -1,20 +1,0 @@
-{
-	"name": "@repo/logger",
-	"version": "0.0.1",
-	"type": "module",
-	"private": true,
-	"imports": {
-		"#*": "./src/*"
-	},
-	"exports": {
-		".": "./src/index.ts"
-	},
-	"scripts": {
-		"dev": "tsc --watch",
-		"check:types": "tsc --noEmit"
-	},
-	"devDependencies": {
-		"@repo/test-preset": "workspace:*",
-		"@repo/typescript-config": "workspace:*"
-	}
-}
