--- packages/logger/tsconfig.json
@@ -1,4 +1,0 @@
-{
-	"extends": "@repo/typescript-config/base.json",
-	"include": ["."]
-}
