--- packages/logger/turbo.json
@@ -1,8 +1,0 @@
-{
-	"extends": ["//"],
-	"tasks": {
-		"test": {
-			"outputs": ["dist/**"]
-		}
-	}
-}
